Added JDBI DAO and mapper.

+++ src/main/java/com/adelsoft/recipie_services/recipie/Recipe.java
@@ -1,12 +1,8 @@
 package com.adelsoft.recipie_services.recipie;
 
-import lombok.Builder;
-import lombok.Getter;
-
 public record Recipe(
-        int id,
+        long id,
         String name,
         String instructions
 ) {
-
 }
